% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Census2016spec-package.R
\docType{package}
\name{Census2016.spec}
\alias{Census2016.spec}
\alias{Census2016.spec-package}
\alias{Census2016.spec-package}
\title{Census2016 specification}
\description{
Census2016 specification
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{package}
