% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colnames.R
\name{check_colnames}
\alias{check_colnames}
\title{Check the column names of tables}
\usage{
check_colnames(data_list, show.progress = TRUE)
}
\arguments{
\item{data_list}{A list of \code{data.tables} to check, or if a string, the name of a package
whose data will be checked.}

\item{show.progress}{Be fun and show a progress bar?}
}
\description{
Check the column names of tables
}
