% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_measures.R
\name{check_measures}
\alias{check_measures}
\title{Check measure columns}
\usage{
check_measures(dt, show.progress = TRUE)
}
\arguments{
\item{dt}{A \code{data.table}.}

\item{show.progress}{Be fun and show a progress bar?}
}
\description{
Check measure columns
}
