% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_country.R
\name{decode_country}
\alias{decode_country}
\title{Decode country to ISO-3166}
\usage{
decode_country(out)
}
\arguments{
\item{out}{A \strong{data.table} containing a column \code{CountryOfBirth}.}
}
\value{
\code{out} with countries decoded to ISO-3166.
}
\description{
Decode country to ISO-3166
}
