% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeRegion.R
\name{decodeRegion}
\alias{decodeRegion}
\title{Decode ASGS/non-ASGS region}
\usage{
decodeRegion(DT)
}
\arguments{
\item{DT}{A \code{data.table} whose first column is a region key will be decoded consistently.}
}
\value{
\code{DT} with the first column coded consistently
}
\description{
Decode ASGS/non-ASGS region
}
