% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var.R
\name{extract_var}
\alias{extract_var}
\title{Extract arbitrary variable from data pack}
\usage{
extract_var(DT, var, ...)
}
\arguments{
\item{DT}{A molten version of the raw data pack files.}

\item{var}{A string, one of the names \code{MEASURE_VARS}, the variable to be extracted.}

\item{...}{Passed to \code{zmatch}.}
}
\description{
Extract arbitrary variable from data pack
}
