% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEASURE_VARS.R
\docType{data}
\name{MEASURE_VARS}
\alias{MEASURE_VARS}
\title{Data dictionary, specification}
\format{An object of class \code{list} of length 100.}
\usage{
MEASURE_VARS
}
\description{
Defines the allowed names, the allowed class for each column names, and the permitted values for each column in the \code{Census2016} series.
}
\keyword{datasets}
