% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mop.R
\name{Mop}
\alias{Mop}
\title{Mop up a data pack file for inclusion a package}
\usage{
Mop(DT, value.name = getOption("Census2016.value.name", "persons"),
  suborder = NULL, dry.run = FALSE, totals = NULL, use.varI = TRUE,
  verbose = FALSE)
}
\arguments{
\item{DT}{A \code{data.table} that separated into the right variables.}

\item{value.name}{The value column in \code{DT}.}

\item{suborder}{If column names are not meant to be ordered alphabetically, this specifies a suborder for columns.}

\item{dry.run}{(logical, default: \code{FALSE}) If \code{TRUE}, the \code{data.table}s are not assigned; the function is run to see if the code will run.}

\item{totals}{A \code{data.table} which specifies the total for each region key.}

\item{use.varI}{(logical) If \code{TRUE}, the default, \code{DT} must contain a column
\code{varI}, the metadata entries, and these \code{varI} will be excluded from subsequent
calls to \code{fread[GTW]}.}

\item{verbose}{Be chatty and report progress (for debugging)?}
}
\value{
Called for its side-effect: assigning to the global environment.
}
\description{
Mop up a data pack file for inclusion a package
}
