% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_check.R
\name{test_check}
\alias{test_check}
\title{Test check}
\usage{
test_check(package, data_list = NULL, show.progress = TRUE)
}
\arguments{
\item{package}{The package as a character vector to check.}

\item{data_list}{Names list of data.tables to check.}

\item{show.progress}{Be fun and show a progress bar?}
}
\description{
Test check
}
