% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colnames.R
\name{check_colnames}
\alias{check_colnames}
\title{Check the column names of tables}
\usage{
check_colnames(data_list, show.progress = TRUE)
}
\arguments{
\item{package}{Quoted name of the package to test.}
}
\description{
Check the column names of tables
}
